requirements.txt
flask
PyYAML
python-pptx
python-docx
fpdf
Pillow
requests
anthropic
gunicorn
Flask-SQLAlchemy
playwright
